Create screenshare context and refactored
