+++ template/fpe-api/components.ts
@@ -1,4 +1,6 @@
 //Precall components
 export * from "../src/components/precall";
 //video call components
-export * from "../src/pages/VideoCall"
+export * from "../src/pages/VideoCall";
+//Share link
+export * from "../src/components/Share";
